Create new submodules for Mapper and RecursivePageTable
