+++ src/structures/paging/mapper/recursive.rs
@@ -1,5 +1,7 @@
 #![cfg(target_arch = "x86_64")]
 
+//! Access the page tables through a recursively mapped level 4 table.
+
 use crate::instructions::tlb;
 use crate::registers::control::Cr3;
 use crate::structures::paging::{
@@ -9,88 +11,8 @@ use crate::structures::paging::{
 };
 use crate::{PhysAddr, VirtAddr};
 use ux::u9;
+use super::*;
 
-/// This type represents a page whose mapping has changed in the page table.
-///
-/// The old mapping might be still cached in the translation lookaside buffer (TLB), so it needs
-/// to be flushed from the TLB before it's accessed. This type is returned from function that
-/// change the mapping of a page to ensure that the TLB flush is not forgotten.
-#[derive(Debug)]
-#[must_use = "Page Table changes must be flushed or ignored."]
-pub struct MapperFlush<S: PageSize>(Page<S>);
-
-impl<S: PageSize> MapperFlush<S> {
-    /// Create a new flush promise
-    fn new(page: Page<S>) -> Self {
-        MapperFlush(page)
-    }
-
-    /// Flush the page from the TLB to ensure that the newest mapping is used.
-    pub fn flush(self) {
-        tlb::flush(self.0.start_address());
-    }
-
-    /// Don't flush the TLB and silence the “must be used” warning.
-    pub fn ignore(self) {}
-}
-
-/// A trait for common page table operations.
-pub trait Mapper<S: PageSize> {
-    /// Creates a new mapping in the page table.
-    ///
-    /// This function might need additional physical frames to create new page tables. These
-    /// frames are allocated from the `allocator` argument. At most three frames are required.
-    ///
-    /// This function is unsafe because the caller must guarantee the following:
-    ///
-    /// - The passed `frame` must be unused, i.e. not used for any other mappings.
-    /// - The passed `frame_allocator` must only yield unused frames.
-    unsafe fn map_to<A>(
-        &mut self,
-        page: Page<S>,
-        frame: PhysFrame<S>,
-        flags: PageTableFlags,
-        frame_allocator: &mut A,
-    ) -> Result<MapperFlush<S>, MapToError>
-    where
-        A: FrameAllocator<Size4KiB>;
-
-    /// Removes a mapping from the page table and returns the frame that used to be mapped.
-    ///
-    /// Note that no page tables or pages are deallocated.
-    fn unmap(&mut self, page: Page<S>) -> Result<(PhysFrame<S>, MapperFlush<S>), UnmapError>;
-
-    /// Updates the flags of an existing mapping.
-    fn update_flags(
-        &mut self,
-        page: Page<S>,
-        flags: PageTableFlags,
-    ) -> Result<MapperFlush<S>, FlagUpdateError>;
-
-    /// Return the frame that the specified page is mapped to.
-    fn translate_page(&self, page: Page<S>) -> Option<PhysFrame<S>>;
-
-    /// Maps the given frame to the virtual page with the same address.
-    ///
-    /// This function is unsafe because the caller must guarantee the following:
-    ///
-    /// - The passed `frame` must be unused, i.e. not used for any other mappings.
-    /// - The passed `frame_allocator` must only yield unused frames.
-    unsafe fn identity_map<A>(
-        &mut self,
-        frame: PhysFrame<S>,
-        flags: PageTableFlags,
-        frame_allocator: &mut A,
-    ) -> Result<MapperFlush<S>, MapToError>
-    where
-        A: FrameAllocator<Size4KiB>,
-        S: PageSize,
-        Self: Mapper<S>,
-    {
-        let page = Page::containing_address(VirtAddr::new(frame.start_address().as_u64()));
-        self.map_to(page, frame, flags, frame_allocator)
-    }
-}
 
 /// A recursive page table is a last level page table with an entry mapped to the table itself.
 ///
@@ -116,38 +38,6 @@ pub struct RecursivePageTable<'a> {
 /// Returned from `RecursivePageTable::new`.
 #[derive(Debug)]
 pub struct NotRecursivelyMapped;
-
-/// This error is returned from `map_to` and similar methods.
-#[derive(Debug)]
-pub enum MapToError {
-    /// An additional frame was needed for the mapping process, but the frame allocator
-    /// returned `None`.
-    FrameAllocationFailed,
-    /// An upper level page table entry has the `HUGE_PAGE` flag set, which means that the
-    /// given page is part of an already mapped huge page.
-    ParentEntryHugePage,
-    /// The given page is already mapped to a physical frame.
-    PageAlreadyMapped,
-}
-
-/// An error indicating that an `unmap` call failed.
-#[derive(Debug)]
-pub enum UnmapError {
-    /// An upper level page table entry has the `HUGE_PAGE` flag set, which means that the
-    /// given page is part of a huge page and can't be freed individually.
-    ParentEntryHugePage,
-    /// The given page is not mapped to a physical frame.
-    PageNotMapped,
-    /// The page table entry for the given page points to an invalid physical address.
-    InvalidFrameAddress(PhysAddr),
-}
-
-/// An error indicating that an `update_flags` call failed.
-#[derive(Debug)]
-pub enum FlagUpdateError {
-    /// The given page is not mapped to a physical frame.
-    PageNotMapped,
-}
 
 impl<'a> RecursivePageTable<'a> {
     /// Creates a new RecursivePageTable from the passed level 4 PageTable.
